src/requirements.txt
# Core API and Server
fastapi
uvicorn[standard]
python-dotenv

# LangChain and its integrations
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai # For talking to the LiteLLM proxy
langchain-ollama # For the Ragas judge model
openai
tiktoken

# Vector Store
faiss-cpu

# Document Loading
pypdf
unstructured
markdown

# LLM Gateway
litellm

# For Ragas evaluation
ragas
pandas
datasets
Pillow

# For load testing
httpx[http2]
numpy
tqdm
matplotlib

# Document processing (hybrid)
docling
pdfplumber
pytesseract
